feat: finilize getters
